Made all of the js features starting with styling
