--- code/src/components/Button.js
@@ -1,5 +1,0 @@
-import React from "react"
-
-export const Button = () => (
-  <button>Push!</button>
-)
