+++ code/src/components/Summary.js
@@ -1,16 +1,12 @@
 import React from "react"
 
 
-
 export const Summary = (props) => {
-  const { song, name, type } = props
+  const { whichSong, name, type } = props
+
   return (
-    <div className="survey-summary">
-
-      <h3 className="">When you wash your hands you sing {props.song}</h3>
-      <h3>you use {props.name} </h3>
-      <h3>in {props.type} form.</h3>
-    </div>
-
+    <section className="survey-summary">
+      <h3>When you wash your hands you sing {whichSong} while you use {name} in {type} form.</h3>
+    </section>
   )
 }
